requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.2.14
dj-rest-auth==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.5.1
typing-extensions==4.12.2
Pillow==9.2.0
django-grappelli==3.0.3
django-taggit==3.0.0
django-simple-captcha==0.5.17
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-filter==22.1